chore: Consolidate namespace-select-filter model into single file
